backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
greenlet==3.0.3
httptools==0.6.4
idna==3.6
imbalanced-learn==0.13.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.2
MarkupSafe==2.1.5
numpy==2.2.4
pandas==2.2.3
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sklearn-compat==0.1.3
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
threadpoolctl==3.6.0
typing_extensions==4.10.0
tzdata==2025.1
uvicorn==0.30.6
